moved files to structure and eat progress functioning
