+++ src/javascripts/components/eat/eat.scss
@@ -3,14 +3,16 @@
   text-align: center; 
 }
 
-#myBar {
-  width: 100%;
-  height: 30px;
-  background-color: rgb(10, 175, 79);
-  text-align: center;
-  line-height: 30px;
-  color: white;
-}
+// #myBar {
+//   width: 100%;
+//   height: 30px;
+//   color: aquamarine;
+//   text-align: center;
+//   line-height: 30px;
+//   color: white;
+
+// }
+
 
 #barWrapper {
   width: 50%;
