+++ src/javascripts/components/fight/fight.scss
