+++ src/javascripts/main.js
@@ -1,11 +1,12 @@
 import '../styles/main.scss';
-import eat from './components/eat';
-import play from './components/play';
-import fight from './components/fight';
-import sleep from './components/sleep';
+import eat from './components/eat/eat';
+import play from './components/play/play';
+import fight from './components/fight/fight';
+import sleep from './components/sleep/sleep';
 
 const init = () => {
   eat.eatPrinter();
+  eat.foodTimer();
   play.playPrinter();
   fight.fightPrinter();
   sleep.sleepPrinter();
