+++ src/styles/main.scss
@@ -8,6 +8,7 @@ html, body {
     display: flex;
     flex-wrap: wrap;
     flex-direction: row;
+    z-index: -1;
 }
   
 #app > .quad {
